added flavour static page into a folder named flavours-static-draft
